+++ src/EventProcessor.php
@@ -68,7 +68,7 @@ class Event
     /**
      * Get a key (serialize) an event object
      *
-     * @return The serialized data
+     * @return string The serialized data
      */
     public function getKey()
     {
@@ -89,12 +89,6 @@ class Event
         $return = [];
         foreach (self::SERIALIZE_MAP as $attribute => $key) {
             $return[$key] = $this->{$attribute};
-        }
-
-        if ($with_metadata) {
-            $return['first_call_timestamp'] = (new \DateTime())->setTimestamp($this->first_call_timestamp)->format('c');
-            $return['last_call_timestamp'] = (new \DateTime())->setTimestamp($this->last_call_timestamp)->format('c');
-            $return['count'] = $this->count;
         }
 
         return $return;
@@ -123,11 +117,12 @@ class Event
  */
 class EventProcessor
 {
-    private static $_creds = null;
-    private static $_last_reported = null;
-    private static $_processing = false;
+    private static ?Credentials $_creds = null;
+    private static ?string $_last_reported = null;
+    private static bool $_processing = false;
+    private static ?string $user_metadata = null;
 
-    const EVENT_TYPE_ENCRYPT = 'encrypt';
+    const EVENT_TYPE_ENCRYPT = 'encrypted';
     const EVENT_TYPE_DECRYPT = 'decrypt';
 
     /**
@@ -202,16 +197,74 @@ class EventProcessor
         }
         ubiq_debug(self::$_creds, 'Not processing; time of ' . self::$_last_reported . ' to now has not exceeded threshold of ' . self::$_creds->config['event_reporting']['flush_interval']);
 
-        if (self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) > self::$_creds->config['event_reporting']['minimum_event_count']
+        if (self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) > self::$_creds->config['event_reporting']['minimum_count']
         ) {
-            ubiq_debug(self::$_creds, 'Processing; count of ' . self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) . ' exceeded threshold of ' . self::$_creds->config['event_reporting']['minimum_event_count']);
+            ubiq_debug(self::$_creds, 'Processing; count of ' . self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) . ' exceeded threshold of ' . self::$_creds->config['event_reporting']['minimum_count']);
 
             return true;
         }
 
-        ubiq_debug(self::$_creds, 'Not processing; count of ' . self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) . ' has not exceeded threshold of ' . self::$_creds->config['event_reporting']['minimum_event_count']);
+        ubiq_debug(self::$_creds, 'Not processing; count of ' . self::$_creds::$cachemanager::getCount(CacheManager::CACHE_TYPE_EVENTS) . ' has not exceeded threshold of ' . self::$_creds->config['event_reporting']['minimum_count']);
 
         return false;
+    }
+
+    /**
+     * Sets user metadata to send with events
+     *
+     * @param string $user_data A valid JSON string less than 1024 characters
+     * 
+     * @return None
+     */
+    public static function addUserDefinedMetadata(string $user_data) {
+        $user_data = json_decode($user_data, TRUE);
+        
+        if (empty($user_data)) {
+            throw new \Exception('User defined metadata must not be null and must be valid JSON');
+
+            return false;
+        }
+        
+        if (strlen($user_data) > 1024) {
+            throw new \Exception('User defined metadata cannot be longer than 1024 characters');
+
+            return false;
+        }
+        
+        $user_data = json_decode($user_data);
+
+        ubiq_debug(self::$_creds, 'Setting user defined metadata to ' . $user_data);
+
+        self::$user_metadata = $user_data;
+    }
+
+    /**
+     * Converts timestamp to formatted datetime with granularity based on configuration
+     *
+     * @param string $timestamp Timestamp to format
+     * 
+     * @return None
+     */
+    private static function formatTimestamp(string $timestamp) {
+        switch (self::$_creds->config['event_reporting']['timestamp_granularity'] ?? "NANOS") {
+            case "MINUTES":
+                return date('c', round($timestamp/60)*60);
+            case "HOURS":
+                return date('c', round($timestamp/60/60)*60*60);
+            case "HALF_DAYS":
+                $dt = date('c', round($timestamp/60/60/24)*60*60*24);
+                if (date('H', $timestamp) > 12) {
+                    $dt += 60*60*12;
+                }
+                return $dt;
+            case "DAYS":
+                return date('c', round($timestamp/60/60/24)*60*60*24);
+            case "NANOS":
+            case "MILLIS":
+            case "SECONDS":
+            default:
+                return date('c', $timestamp);
+        }
     }
 
     /**
@@ -238,13 +291,25 @@ class EventProcessor
         // format for reporting
         $events = [];
         foreach ($cached_events as $cached_event) {
-            $event = $cached_event->getMappedArray(true);
+            $event = $cached_event->getMappedArray();
 
             $event['product'] = \Ubiq\LIBRARY;
             $event['product_version'] = \Ubiq\VERSION;
             $event['user-agent'] = \Ubiq\LIBRARY . '/' . \Ubiq\VERSION;
             $event['api_version'] = \Ubiq\API_VERSION;
 
+            // set user metadata
+            if (!empty(self::$user_metadata)) {
+                $event['user_defined'] = self::$user_metadata;
+            }
+    
+            // set count
+            $event['count'] = $cached_event->count;
+
+            // set time stamp granularity
+            $event['first_call_timestamp'] = self::formatTimestamp($cached_event->first_call_timestamp);
+            $event['last_call_timestamp'] = (new \DateTime())->setTimestamp($cached_event->last_call_timestamp)->format('c');
+    
             $events[] = $event;
         }
         $events = ['usage' => $events];
@@ -265,18 +330,27 @@ class EventProcessor
             self::$_creds->getSapi()
         );
 
-        if ($async) {
-            $resp = $http->postAsync(
-                self::$_creds->getHost() . '/api/v3/tracking/events',
-                json_encode($events),
-                'application/json'
-            );
-        } else {
-            $resp = $http->post(
-                self::$_creds->getHost() . '/api/v3/tracking/events',
-                json_encode($events),
-                'application/json'
-            );
+        try {
+            if ($async) {
+                $resp = $http->postAsync(
+                    self::$_creds->getHost() . '/api/v3/tracking/events',
+                    json_encode($events),
+                    'application/json'
+                );
+            } else {
+                $resp = $http->post(
+                    self::$_creds->getHost() . '/api/v3/tracking/events',
+                    json_encode($events),
+                    'application/json'
+                );
+            }
+        }
+        catch (\Exception $e) {
+            if (self::$_creds->config['event_reporting']['trap_exceptions'] ?? false == true) {
+                throw $e;
+            }
+
+            ubiq_debug(self::$_creds, 'Ignored exception from event reporting ' . $e->getMessage());
         }
 
         ubiq_debug(self::$_creds, 'Processed ' . sizeof($events['usage']) . ' events');