+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.0 - 6/27/2023
+* Added user defined metadata to events
+* Configurable error handling for event reporting
+* Time granularity for event reporting
+* Add support for cache TTL
+* Update config structure
+
+
 ## 0.1.0 - 6/27/2023
 * Added key caching for unstructured data
 * Added cache manager singleton for key and event log (and eventually dataset) caching